Categorizer/requirements.txt
httpx>=0.24.1
twikit>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
pytest>=7.4.0
python-dateutil>=2.8.2 